backend: youtube: implement commentModerated hook.
