+++ packages/backend-api/src/integrations/youtube/comments.ts
@@ -18,11 +18,19 @@ import { OAuth2Client } from 'google-auth-library';
 import { google } from 'googleapis';
 
 import { logger } from '@conversationai/moderator-backend-core';
-import { IUserInstance, MODERATION_ACTION_ACCEPT, MODERATION_ACTION_DEFER } from '@conversationai/moderator-backend-core';
+import {
+  ICommentInstance,
+  IDecisionInstance,
+  IUserInstance,
+  MODERATION_ACTION_ACCEPT,
+  MODERATION_ACTION_DEFER,
+} from '@conversationai/moderator-backend-core';
 
+import { foreachPendingDecision, markDecisionExecuted } from '../decisions';
 import { for_each_active_channel } from './channels';
-import { foreachPendingDecision, mapCommentThreadToComments, markDecisionExecuted } from './objectmap';
+import { mapCommentThreadToComments } from './objectmap';
 import { get_article_id_from_youtube_id } from './videos';
+import { } from '@conversationai/moderator-backend-core/src';
 
 const service = google.youtube('v3');
 
@@ -110,33 +118,41 @@ export async function sync_comment_threads(
   });
 }
 
+export async function implement_moderation_decision(
+  auth: OAuth2Client,
+  comment: ICommentInstance,
+  decision: IDecisionInstance,
+) {
+  const sourceId = comment.get('sourceId') as string;
+  const status = decision.get('status');
+
+  if (status === MODERATION_ACTION_DEFER) {
+    logger.info('Not syncing comment %s:%s - in deferred state', comment.id, sourceId);
+    markDecisionExecuted(decision);
+    return;
+  }
+
+  const moderationStatus = (status === MODERATION_ACTION_ACCEPT) ? 'published' : 'rejected';
+  logger.info('Syncing comment %s:%s to %s (%s) ', comment.id, sourceId, moderationStatus, decision.id);
+  service.comments.setModerationStatus({
+      auth: auth,
+      id: sourceId,
+      moderationStatus: moderationStatus,
+    },
+    (err) => {
+      if (err) {
+        logger.error(`Google API returned an error for comment ${comment.id}: ` + err);
+        return;
+      }
+      markDecisionExecuted(decision);
+    });
+}
+
 export async function implement_moderation_decisions(
   owner: IUserInstance,
   auth: OAuth2Client,
 ) {
   await foreachPendingDecision(owner, async (decision, comment) => {
-    const sourceId = comment.get('sourceId') as string;
-    const status = decision.get('status');
-
-    if (status === MODERATION_ACTION_DEFER) {
-      logger.info('Not syncing comment %s:%s - in deferred state', comment.id, sourceId);
-      markDecisionExecuted(decision);
-      return;
-    }
-
-    const moderationStatus = (status === MODERATION_ACTION_ACCEPT) ? 'published' : 'rejected';
-    logger.info('Syncing comment %s:%s to %s (%s) ', comment.id, sourceId, moderationStatus, decision.id);
-    service.comments.setModerationStatus({
-        auth: auth,
-        id: sourceId,
-        moderationStatus: moderationStatus,
-      },
-      (err) => {
-        if (err) {
-          logger.error(`Google API returned an error for comment ${comment.id}: ` + err);
-          return;
-        }
-        markDecisionExecuted(decision);
-      });
+    await implement_moderation_decision(auth, comment, decision);
   });
 }
