+++ packages/backend-api/src/integrations/youtube/task.ts
@@ -16,7 +16,7 @@ limitations under the License.
 
 import { for_all_youtube_users } from './authenticate';
 import { sync_channels } from './channels';
-import { sync_comment_threads } from './comments';
+import {implement_moderation_decisions, sync_comment_threads} from './comments';
 import { clearError } from './objectmap';
 import { sync_known_videos } from './videos';
 
@@ -31,6 +31,7 @@ export async function syncYoutubeTask(tick: number) {
         await clearError(owner);
         await sync_channels(owner, auth);
         await sync_known_videos(owner, auth);
+        await implement_moderation_decisions(owner, auth);
       }
       await sync_comment_threads(owner, auth, false);
     });
